Converted to ESM module format from CommonJS.
The motivation for doing so is to be able to use the current version of Chai for
testing which only supports ESM.

The changes touch many files, but they are small, and there are very few real
code changes.

1. A big requirement is that all relative file import statements MUST have the
   ".js" extension added.

2. In interpreter/src/index.ts and interpreter/src/indexGSuite.ts, the export
   for DevEnvironment must be "type DevEnvironment" since it is a type.

3. In cli/bin/index.js, the "require" must be changed to an "import"-"export".

4. In cli/src/commandLine.ts, the imports of commandLineArgs and commandLineUsage
   needed to be changed from "import * as" to just "import".

5. For transpiling the TypeScript code to JavaScript, the key changes for ESM
   instead of CommonJS support are in the compiler options in tsconfig.base.json
   to specify:
   	 "target": "es2020" (instead of "es6"),
   	 "module": "Node16" (instead of "commonjs"),
   	 "moduleResolution": "Node16" (instead of the default of "node")

6. Additionally, the package.json file for each of the packages needs to specify:
     "type": "module"

5. The deploy.sh script in sheets_addon required the biggest changes,
   browserify+tsify doesn't work because browserify only supports CommonJS. I was
   unsuccessful in my attempts to solve the problem using browserify+tsify+esmify.
   I eventually discovered that I could us rollup to create a single file in
   CommonJS module format from the compiled ESM JS for gramble, and then use
   browserify to finish the job of creating a code file that is compatible with
   Google AppsScript. The tsconfig.json file is no longer needed for this package,
   but we add a rollup.config.js instead.

6. A small code change was needed in the tests package because ESM does not
   have a "module" object (with a filename property). Instead of module, ESM
   provides the module filename via "import.meta.filename". So, at the top of
   each test file, we add:
     const module = import.meta
   Then the code works, as is, after adjusting a couple parameter types to now be
     NodeModule | ImportMeta

7. The other key change in the tests, is to run mocha with ts-node/esm as the
   loader (and "ts-node/register" is no longer required). The easy way to do
   this in the package.json is to add a mocha section to specify the mocha option:
     "loader": "ts-node/esm"
   Then these common options don't need to be specified for each NPM test* script.

8. Also, for running the tests in the Mocha Test Explorer in VSCode, we need to
   adjust the mochaExplorer settings to add ts-node/esm as the loader:
     "mochaExplorer.nodeArgv": ["--loader=ts-node/esm"]
   and to remove the "mochaExplorer.require": "ts-node/register" line.

9. Finally, we commented out some code in the property tests that was causing
   compilation errors. The first was in tests/property/reduceGrammar.ts: there
   is actually a bug here because the last changes for QualifiedGrammer are
   incomplete; the current code assumed that QualifiedGrammar is a
   CollectionGrammar and resulted in a type error. For now, the offending line
   is commented out. The second issue was in testPropertyUtil.ts where an
   unnecessary line was now being flagged as such, so we commented it out.

+++ packages/sheets_addon/deploy.sh
@@ -27,16 +27,22 @@ PATH="$PWD"/../../node_modules/.bin:$PATH
 PROJ_NAME=$1
 TARGET_DIR=../../deployments/$PROJ_NAME
 
-# transpile Gramble to javascript and wrap it as HTML
-echo "Transpiling Gramble to JavaScript..." >&2
-browserify ../interpreter/src/indexGSuite.ts -p tsify -s gramble -o gramble.js
+# Create a single Gramble javascript file and wrap it as HTML
+echo "Rolling up Gramble into a single file..." >&2
+rollup ../interpreter/dist/indexGSuite.js -c --file gramble1.js --format cjs
+echo "Browserifying gramble.js..." >&2
+browserify --extension=.js -s gramble -o gramble.js gramble1.js
+if [[ ! -f gramble.js ]]; then
+    echo "Error: gramble.js file does not exist" >&2
+    exit 1
+fi
 cat <(echo '<script>') gramble.js <(echo '</script>') >> grambleWrapped.html
 
 # make sure the target directory exists, and copy the project files to it
 echo "Copying project files to '$PWD'/$TARGET_DIR" >&2
 mkdir -p $TARGET_DIR
 cp Code.js gramble.js grambleWrapped.html sidebar.html style.html appsscript.json $TARGET_DIR
-rm gramble.js grambleWrapped.html 
+rm gramble1.js gramble.js grambleWrapped.html 
 
 cd $TARGET_DIR
